pyproject.toml
[project]
name = "summaries"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.21", # Required by hooks/relationship.py
    "jieba>=0.42.1", # 搜索中文
    "mdx-truly-sane-lists>=1.3",
    "mkdocs-material>=9.6.22",
    "mkdocs-redirects>=1.2.2",
    "mkdocs-rss-plugin>=1.17.4",
    "typst>=0.13.7", # Required by hooks/typst_math.py
]

[tool.ruff.lint]
extend-select = [
    "U",
    "TYP",
    "F",
    "I",
]
